+++ Task/app_uart_parser.h
@@ -56,6 +56,7 @@ typedef enum {
     HCMD_VALVE_ON,            // 开启电磁阀
     HCMD_VALVE_OFF,           // 关闭电磁阀
     HCMD_LIGHT,               // 切换下补光灯
+    HCMD_CALIB_ENC,           // P2编码器比例标定
 
     /* 未知命令 */
     HCMD_UNKNOWN